Moving DataContractUtils to common client
